requirements.txt
Flask==2.3.3
WeasyPrint==57.0
Werkzeug==2.3.7
python-dotenv 
requests
youtube_transcript_api
opencv-python
yt-dlp
google-genai
pdfkit
flask-cors